fix: Resolve final 6 clippy errors for v0.4.2 CI/CD Health
Fixed remaining pedantic lints in test code:
- scheduler_verification.rs: Remove unnecessary u64 cast
- message.rs: Replace no-op match with simple ignore
- standby.rs: Allow float_cmp in test assertions
- simulation.rs: Allow large_enum_variant for SimEvent (Message is 416 bytes)
- Add allow attributes to test files for pedantic lints

All clippy checks now pass with -D warnings.

Part of v0.4.2 CI/CD Health milestone.

+++ crates/kimberlite-sim/src/scheduler_verification.rs
@@ -486,7 +486,7 @@ mod tests {
 
         // Record 200 decisions
         for i in 0..200 {
-            tracker.record_schedule((i % 2) as u64, i * 1000, vec![0, 1]);
+            tracker.record_schedule(i % 2, i * 1000, vec![0, 1]);
         }
 
         // Should only keep 100
